chore: split large modules in amaru-tui in sub-components.
