+++ crates/amaru-tui/src/ui/components/card.rs
@@ -1,0 +1,36 @@
+// Copyright 2026 PRAGMA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use ratatui::{
+    Frame,
+    layout::Rect,
+    text::Line,
+    widgets::{Block, Borders, Paragraph, Wrap},
+};
+
+use super::super::theme::{block_title, border_secondary};
+use crate::model::InteractionMode;
+
+pub(in crate::ui) fn render_card(
+    frame: &mut Frame<'_>,
+    area: Rect,
+    title: &str,
+    lines: Vec<Line<'static>>,
+    mode: InteractionMode,
+) {
+    let paragraph = Paragraph::new(lines).wrap(Wrap { trim: true }).block(
+        Block::default().title(block_title(mode, title)).borders(Borders::ALL).border_style(border_secondary(mode)),
+    );
+    frame.render_widget(paragraph, area);
+}
