+++ crates/amaru-tui/src/ui/components/mod.rs
@@ -1,0 +1,26 @@
+// Copyright 2026 PRAGMA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod card;
+mod config_sections;
+mod epoch_progress;
+mod logs;
+mod peers;
+mod proposals;
+mod series_card;
+
+pub(super) use self::{
+    card::render_card, config_sections::render_section_groups, epoch_progress::render_epoch_progress,
+    logs::render_logs, peers::render_peers_table, proposals::render_proposals_table, series_card::render_series_card,
+};
